extract logic for finding SSH environment into module
